extract processOptions and added unit tests
